# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Verifyflow API
  version: 0.5.0
  description: |
    Public HTTP contract for the registered health, authentication, account, rental, payment,
    messaging, webhook, and admin routes. All customer request bodies are JSON. Responses include an `X-Request-ID` header.
    Access tokens are short-lived JWTs; refresh tokens are opaque, single-use tokens that rotate
    on refresh.

    Rental routes require a verified email address in addition to a valid access token. Admin
    routes require the signed-in user's role to be `admin`; every account starts as `user`.

    Twilio, Vonage, and Telnyx are supported behind a provider-neutral interface. Inbound provider
    callbacks require their vendor signature and are persisted idempotently.
servers:
  - url: /
    description: Same origin as this documentation page
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register and sign in
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created and signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          $ref: '#/components/responses/InvalidJSON'
        '409':
          description: Email is already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: email_taken
                  message: An account with this email already exists.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Sign in
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Signed in; a new independent session is created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          $ref: '#/components/responses/InvalidJSON'
        '401':
          description: Email or password is incorrect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_credentials
                  message: Invalid email or password.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate refresh token
      description: Exchanges a refresh token for a new access token and refresh token. Serialize refresh calls.
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens rotated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/InvalidJSON'
        '401':
          description: Refresh token is invalid, expired, revoked, or reused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_refresh_token
                  message: Refresh token is invalid or expired. Please sign in again.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke a refresh-token session
      description: Idempotent. The current access token remains valid until it expires.
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Session revoked, or token was unknown or expired.
        '400':
          $ref: '#/components/responses/InvalidJSON'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password-reset link
      description: Always returns the same accepted response for a well-formed email to prevent account enumeration.
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '202':
          description: Request accepted; no account-existence information is disclosed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: If an account exists for that email, a password reset link has been sent.
        '400':
          $ref: '#/components/responses/InvalidJSON'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Set a password using a reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password changed; all refresh sessions are revoked.
        '400':
          description: Invalid JSON or reset token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidJson:
                  $ref: '#/components/examples/InvalidJSON'
                invalidToken:
                  value:
                    error:
                      code: invalid_reset_token
                      message: This reset link is invalid or has expired. Please request a new one.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/auth/verify-email:
    post:
      tags: [Authentication]
      summary: Verify an email address
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '204':
          description: Email address verified.
        '400':
          description: Invalid JSON or verification token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidJson:
                  $ref: '#/components/examples/InvalidJSON'
                invalidToken:
                  value:
                    error:
                      code: invalid_verification_token
                      message: This verification link is invalid or has expired. Request a new one from your account.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/auth/resend-verification:
    post:
      tags: [Authentication]
      summary: Request a new verification email
      operationId: resendVerification
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Email address is already verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Your email is already verified.
        '202':
          description: Verification email queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Verification email sent.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/auth/change-password:
    post:
      tags: [Authentication]
      summary: Change the signed-in user's password
      description: Revokes all refresh sessions and returns replacement tokens for this device.
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed; replacement tokens returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid JSON or account has no password set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidJson:
                  $ref: '#/components/examples/InvalidJSON'
                passwordNotSet:
                  value:
                    error:
                      code: password_not_set
                      message: This account has no password yet. Use "forgot password" to set one.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Concurrent password change conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: conflict
                  message: Your password was changed by another request. Please sign in again.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/me:
    get:
      tags: [Account]
      summary: Get the authenticated user
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Authenticated user's profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/numbers:
    get:
      tags: [Rentals]
      summary: Search available numbers
      description: Lists provider numbers currently AVAILABLE for rental, optionally filtered by type and capability.
      operationId: searchNumbers
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [Local, TollFree, Mobile]
        - name: sms
          in: query
          description: When true, only return numbers with SMS capability.
          schema:
            type: boolean
        - name: mms
          in: query
          description: When true, only return numbers with MMS capability.
          schema:
            type: boolean
        - name: voice
          in: query
          description: When true, only return numbers with voice capability.
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: Opaque pagination cursor from a previous page's next_cursor.
          schema:
            type: string
      responses:
        '200':
          description: Page of available numbers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NumbersPage'
        '400':
          description: Invalid type or limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_request
                  message: Request is missing or has invalid fields.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmailNotVerified'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/rental-plans:
    get:
      tags: [Rentals]
      summary: List active rental plans
      operationId: listRentalPlans
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Active rental plans.
          content:
            application/json:
              schema:
                type: object
                required: [plans]
                properties:
                  plans:
                    type: array
                    items:
                      $ref: '#/components/schemas/Plan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmailNotVerified'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/rentals:
    post:
      tags: [Rentals]
      summary: Reserve a number (checkout)
      description: |
        Reserves a provider number under a rental plan and creates a server-priced PENDING
        order. Reusing the same Idempotency-Key with the same request returns the original
        order instead of creating a new one; reusing it with a different number or plan is a
        conflict. The reservation expires after a server-configured TTL if never paid.
      operationId: reserveNumber
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 200
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveRequest'
      responses:
        '201':
          description: Reservation and PENDING order created (or the original order, if this Idempotency-Key was already used).
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Invalid JSON, missing fields, or missing Idempotency-Key header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidJson:
                  $ref: '#/components/examples/InvalidJSON'
                missingIdempotencyKey:
                  value:
                    error:
                      code: missing_idempotency_key
                      message: Idempotency-Key header is required.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmailNotVerified'
        '404':
          description: The requested number was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: number_not_found
                  message: The requested number was not found.
        '409':
          description: Number/plan is no longer available, or this Idempotency-Key was already used with a different request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                numberUnavailable:
                  value:
                    error:
                      code: number_unavailable
                      message: The requested number is no longer available.
                planUnavailable:
                  value:
                    error:
                      code: plan_unavailable
                      message: The requested rental plan is unavailable.
                idempotencyConflict:
                  value:
                    error:
                      code: idempotency_conflict
                      message: This idempotency key was already used with a different request.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/orders/{id}:
    get:
      tags: [Rentals]
      summary: Get an order owned by the signed-in user
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathID'
      responses:
        '200':
          description: The order.
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/EmailNotVerified'
        '404':
          description: No such order, or it belongs to another user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: not_found
                  message: The requested order was not found.
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/admin/users/{id}/role:
    patch:
      tags: [Admin]
      summary: Change another user's role
      description: |
        Admin-only. Refuses to demote the last remaining admin (returns 409 last_admin), so an
        admin can never lock every admin out of the system.
      operationId: setUserRole
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRoleRequest'
      responses:
        '200':
          description: Role updated (or unchanged, if it already matched).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          description: Invalid JSON or unknown role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidJson:
                  $ref: '#/components/examples/InvalidJSON'
                invalidRole:
                  value:
                    error:
                      code: invalid_role
                      message: 'role must be one of: user, admin.'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No such user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: user_not_found
                  message: User not found.
        '409':
          description: The target is the last remaining admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: last_admin
                  message: Cannot remove the last remaining admin.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
...
  /api/v1/admin/provider-configs/{id}:
    patch:
      tags: [Admin]
      summary: Enable or disable a provider config
      operationId: setProviderConfigEnabled
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderConfigEnabledRequest'
      responses:
        '200':
          description: Updated provider config.
          content:
            application/json:
              schema:
                type: object
                required: [provider_config]
                properties:
                  provider_config:
                    $ref: '#/components/schemas/ProviderConfig'
        '400':
          $ref: '#/components/responses/InvalidJSON'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No such provider config.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: not_found
                  message: The requested provider config was not found.
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/Timeout'
  /api/v1/my-numbers:
    get:
      tags: [Messaging]
      summary: List the authenticated customer's rented numbers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer number list.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/webhooks/telephony/{providerConfigID}/inbound:
    post:
      tags: [Webhooks]
      summary: Receive a signed inbound SMS callback
      parameters:
        - in: path
          name: providerConfigID
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Event accepted or already processed. }
        '400': { description: Malformed provider event. }
        '401': { description: Invalid provider signature. }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [full_name, email, password]
      additionalProperties: false
      properties:
        full_name:
          type: string
          maxLength: 100
          description: Required; trimmed and must not contain control characters.
          example: Feranmi Oresajo
        email:
          type: string
          format: email
          maxLength: 254
          example: fer@example.com
        password:
          type: string
          format: password
          minLength: 10
          maxLength: 128
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...